requirements.txt
fastembed=0.6.1
google-generativeai=0.8.5
uuid=1.30
pydantic=2.10.4
qdrantclient=1.14.2
pandas=2.2.3
dotenv=0.9.9